+++ packages/cookie-banner/src/detectors/detector.ts
